+++ packages/Bio/src/tests/monomer-libraries-tests.ts
@@ -27,19 +27,19 @@ category('monomerLibraries', () => {
   test('default', async () => {
     // Clear settings to test default
     await setUserLibSettings({exclude: [], explicit: []});
-    await monomerLibHelper.loadLibraries(true); // test defaultLib
+    await monomerLibHelper.loadMonomerLib(true); // test defaultLib
 
     // Currently default monomer lib set is of all files at LIB_PATH (at least HELMCoreLibrary.json)
-    const currentMonomerLib = monomerLibHelper.getBioLib();
+    const currentMonomerLib = monomerLibHelper.getMonomerLib();
     expect(currentMonomerLib.getPolymerTypes().length > 0, true);
   });
 
   test('forTests', async () => {
     await setUserLibSettingsForTests();
-    await monomerLibHelper.loadLibraries(true); // test defaultLib
+    await monomerLibHelper.loadMonomerLib(true); // test defaultLib
 
     // Currently default monomer lib set is of all files at LIB_PATH (at least HELMCoreLibrary.json)
-    const currentMonomerLib = monomerLibHelper.getBioLib();
+    const currentMonomerLib = monomerLibHelper.getMonomerLib();
     // HELMCoreLibrary.json checks
     expect(currentMonomerLib.getPolymerTypes().length, 2);
     expect(currentMonomerLib.getMonomerSymbolsByType('PEPTIDE').length, 322);
@@ -59,8 +59,8 @@ category('monomerLibraries', () => {
     libSettings.explicit = [];
     await setUserLibSettings(libSettings);
 
-    await monomerLibHelper.loadLibraries(true);
-    const currentMonomerLib = monomerLibHelper.getBioLib();
+    await monomerLibHelper.loadMonomerLib(true);
+    const currentMonomerLib = monomerLibHelper.getMonomerLib();
     const polymerTypes = currentMonomerLib.getPolymerTypes();
     expect(polymerTypes.length === 0, true);
   });